requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
mediapipe>=0.10.0
onnxruntime>=1.15.0
